fix pertag to work with combo
Comboview broke pertag togglebar. These changes fix a rendering bug that
prevented the abr from being hidden. The bar appeared frozen and didn't
reflect the actual status of the bar. Now when disabling the bar, it
only disables it for that tag just like it did before the combo patch.