Violet-ify and fix theme/combo patches
This adds all of my changes that I like:
- window border pixel = 2 (not 1)
- snap pixel = 8 (not 32)
- multiple themes that also change the background (see setbg())
  - tokyonight-moon, kanagawa-wave, catppuccin-mocha, gruvbox8
  - switch between them with Win-Alt-s
- mfact = 0.5 (not 0.55)
- don't respect resize hints
- use Win key as mod (not Alt)
- Win-Return (not Win-Shift-Return) = open terminal
- Win-Alt-u = min brightness
- Win-Alt-i = decrease brightness
- Win-Alt-o = increase brightness
- Win-Alt-p = max brightness
- Win-Alt-h = toggle mute
- Win-Alt-j = decrease volume
- Win-Alt-k = increase volume
- Win-Alt-l = display notification with volume info

With the setbg() function implemented and added to setup() and
setscheme() functions, the background automatically updates. No need to
call `xsetroot -solid ...` anymore. It updates with the theme as well.

The combo patch is effectively removed, but now when you `tag()` another
tag/workspace, it will not change the previously selected tags. This
means you can (eg) go to tag 1 and also bring in tag 2 and then go to
tag 3 and bring in tag 4 and then you can Win-Tab between these two sets
of tags (1+2 and 3+4). No need for a combo patch!

TODO: figure out how to make setscheme() update the window borders, as
well.